Rename to stock_vertical_lift
Preparing the module to be a generic core for supporting Vertical Lifts
/ Carousels. The Kardex one willbe a submodule.
