--- stock_kardex/data/stock_location.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-
-  <record id="stock_location_kardex" model="stock.location">
-    <field name="name">Kardex View</field>
-    <!-- Beware, must not be view, even if we should have stock here.
-         Because stock moves targetting views are not included in the
-         inventory at date, and we want to be able to use "Kardex View"
-         as destination (the final destination will be a "cell" in a kardex).
-    -->
-    <field name="location_id" ref="stock.stock_location_stock"/>
-    <field name="usage">internal</field>
-    <field name="kardex_location" eval="True"/>
-    <field name="company_id"></field>
-  </record>
-
-</odoo>
