--- stock_kardex/demo/stock_kardex_demo.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo noupdate="1">
-
-  <record id="stock_kardex_demo_shuttle_1" model="stock.kardex">
-    <field name="name">Shuttle 1</field>
-    <field name="location_id" ref="stock_location_kardex_demo_shuttle_1"/>
-    <field name="address">10.0.30.1</field>
-    <field name="simulate" eval="True"/>
-    <field name="mode">pick</field>
-  </record>
-
-  <record id="stock_kardex_demo_shuttle_2" model="stock.kardex">
-    <field name="name">Shuttle 2</field>
-    <field name="location_id" ref="stock_location_kardex_demo_shuttle_2"/>
-    <field name="address">10.0.30.2</field>
-    <field name="simulate" eval="True"/>
-    <field name="mode">pick</field>
-  </record>
-
-  <record id="stock_kardex_demo_shuttle_3" model="stock.kardex">
-    <field name="name">Shuttle 3</field>
-    <field name="location_id" ref="stock_location_kardex_demo_shuttle_3"/>
-    <field name="address">10.0.30.3</field>
-    <field name="simulate" eval="True"/>
-    <field name="mode">pick</field>
-  </record>
-
-</odoo>
