--- stock_kardex/models/__init__.py
@@ -1,5 +1,0 @@
-from . import stock_kardex
-from . import stock_kardex_tray_type
-from . import stock_location
-from . import stock_move
-from . import stock_quant
