--- stock_kardex/static/src/scss/kardex_view.scss
@@ -1,129 +1,0 @@
-.o_field_kardex_tray_matrix {
-    background-color: #eeeeee;
-    border: 2px #000000 solid;
-}
-
-.o_web_client.o_fullscreen {
-    $o-kardex-padding: $o-horizontal-padding;
-
-    .o_form_view.o_kardex {
-        display: flex;
-        flex-flow: column nowrap;
-        padding: 0;
-
-        font-size: 16px;
-
-        @include media-breakpoint-up(xl) {
-            font-size: 18px;
-        }
-
-        .btn {
-            font-size: 1em;
-            padding: 1em;
-            margin: 0 5px;
-        }
-
-        .o_kardex_header {
-            display: flex;
-            flex-flow: row wrap;
-            padding: $o-kardex-padding;
-        }
-
-        .o_kardex_header_content {
-            display: flex;
-            flex-flow: row nowrap;
-            font-size: 2.0em;
-            flex: 1 0 auto;
-            align-items: center;
-            width: 33%;
-
-            &.o_kardex_header_right {
-                justify-content: flex-end;
-            }
-        }
-
-        .o_kardex_actions {
-            display: flex;
-            flex-flow: row nowrap;
-            font-size: 1.2em;
-            padding: $o-kardex-padding * 0.5;
-        }
-
-        .o_kardex_operation {
-            text-align: center;
-            font-size: 2.5em;
-            padding: 0.5em;
-            color: #ffffff;
-        }
-
-        .o_kardex_content {
-            display: flex;
-            flex-flow: row nowrap;
-            flex: 1 0 auto;
-            align-items: center;
-
-            &.o_kardex_content_right {
-                justify-content: flex-end;
-            }
-        }
-
-        .o_kardex_data {
-            display: flex;
-            flex-flow: row wrap;
-            padding: $o-kardex-padding * 0.5;
-
-            .o_kardex_data_content {
-                flex-flow: row nowrap;
-                font-size: 1.2em;
-                flex: 1 0 auto;
-                align-items: center;
-                width: 50%;
-
-                &.o_kardex_tray {
-                    display: flex;
-                    justify-content: flex-end;
-
-                    .o_group {
-                        display: block;
-                    }
-                }
-
-                .o_field_kardex_tray_matrix {
-                    width: 450px;
-                }
-            }
-
-            .kardex_highlight {
-                padding: 6px;
-                border-radius: 10px;
-            }
-        }
-
-    }
-
-    .o_kardex_menu {
-        .btn {
-            margin-bottom: $o-kardex-padding;
-            padding: 1em;
-            font-size: 2em;
-            text-transform: uppercase;
-        }
-
-        .o_kardex_exit {
-            text-align: center;
-        }
-    }
-
-    .o_kardex_manual_barcode {
-        .o_field_char {
-            padding: 1em;
-            font-size: 2em;
-        }
-
-        .btn {
-            padding: 1em;
-            font-size: 2em;
-            text-transform: uppercase;
-        }
-    }
-}
