--- stock_kardex/templates/kardex_screen.xml
@@ -1,17 +1,0 @@
-<odoo>
-
-<template id="packagings">
-  <ul class="o_kardex_packaging list-unstyled">
-    <t t-foreach="packagings" t-as="packaging">
-      <li>
-        <span>1</span>
-        <span class="packaging_name" itemprop="name" t-esc="packaging['name']"/>
-        <span>: </span>
-        <span class="packaging_qty" itemprop="qty" t-esc="packaging['qty']"/>
-        <span class="packaging_unit" itemprop="unit" t-esc="packaging['unit']"/>
-      </li>
-    </t>
-  </ul>
-</template>
-
-</odoo>
