--- stock_kardex/tests/__init__.py
@@ -1,3 +1,0 @@
-from . import test_kardex_location
-from . import test_kardex_tray_type
-from . import test_kardex
