--- stock_kardex/views/stock_location_views.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-
-  <record id="view_location_form_kardex" model="ir.ui.view">
-    <field name="name">stock.location.form.kardex</field>
-    <field name="model">stock.location</field>
-    <field name="inherit_id" ref="stock.view_location_form"/>
-    <field name="arch" type="xml">
-      <field name="return_location" position="after">
-        <field name="kardex_location"
-               attrs="{'invisible': [('kardex_kind', '!=', False), ('kardex_kind', '!=', 'view')]}"
-               />
-      </field>
-      <group name="localization" position="after">
-        <group string="Kardex" name="kardex"
-               attrs="{'invisible': [('kardex_kind', 'not in', ('shuttle', 'tray', 'cell'))]}">
-          <field name="kardex_kind"/>
-          <field name="kardex_tray_type_id"
-                 attrs="{'invisible': [('kardex_kind', '!=', 'tray')], 'required': [('kardex_kind', '=', 'tray')]}"/>
-          <field name="tray_matrix"
-                 widget="kardex_tray_matrix"
-                 attrs="{'invisible': [('kardex_kind', 'not in', ('tray', 'cell'))]}"
-                 options="{'click_action': 'action_tray_matrix_click'}"
-                 />
-        </group>
-      </group>
-
-      <field name="posx" position="attributes">
-        <attribute name="attrs">{'readonly': [('kardex_kind', '=', 'cell')]}</attribute>
-      </field>
-      <field name="posy" position="attributes">
-        <attribute name="attrs">{'readonly': [('kardex_kind', '=', 'cell')]}</attribute>
-      </field>
-    </field>
-  </record>
-
-  <record id="view_location_search_kardex" model="ir.ui.view">
-    <field name="name">stock.location.search.kardex</field>
-    <field name="model">stock.location</field>
-    <field name="inherit_id" ref="stock.view_location_search"/>
-    <field name="arch" type="xml">
-      <field name="location_id" position="after">
-        <separator/>
-        <field name="kardex_kind"/>
-        <field name="kardex_tray_type_id"/>
-      </field>
-    </field>
-  </record>
-
-</odoo>
