+++ stock_vertical_lift/data/stock_location.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+
+  <record id="stock_location_vertical_lift" model="stock.location">
+    <field name="name">Vertical Lift</field>
+    <!-- Beware, must not be view, even if we should have stock here.
+         Because stock moves targeting views are not included in the
+         inventory at date, and we want to be able to use "Vertical Lift View"
+         as destination (the final destination will be a "cell" in a Shuttle).
+    -->
+    <field name="location_id" ref="stock.stock_location_stock"/>
+    <field name="usage">internal</field>
+    <field name="vertical_lift_location" eval="True"/>
+    <field name="company_id"></field>
+  </record>
+
+</odoo>
