+++ stock_vertical_lift/demo/stock_inventory_demo.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo noupdate="1">
+
+  <record id="stock_inventory_vertical_lift_0" model="stock.inventory">
+    <field name="name">Starting Vertical Lift Inventory</field>
+  </record>
+
+  <record id="stock_inventory_vertical_lift_line_1" model="stock.inventory.line">
+    <field name="product_id" ref="product_running_socks"/>
+    <field name="product_uom_id" ref="uom.product_uom_unit"/>
+    <field name="inventory_id" ref="stock_inventory_vertical_lift_0"/>
+    <field name="product_qty">30.0</field>
+    <field name="location_id" ref="stock_location_vertical_lift_demo_tray_1b_x3y2"/>
+  </record>
+
+  <function model="stock.inventory" name="action_validate">
+    <function eval="[[('state','=','draft'),('id', '=', ref('stock_vertical_lift.stock_inventory_vertical_lift_0'))]]" model="stock.inventory" name="search"/>
+  </function>
+
+</odoo>
