+++ stock_vertical_lift/models/__init__.py
@@ -1,0 +1,5 @@
+from . import vertical_lift_shuttle
+from . import vertical_lift_tray_type
+from . import stock_location
+from . import stock_move
+from . import stock_quant
