+++ stock_vertical_lift/tests/__init__.py
@@ -1,0 +1,3 @@
+from . import test_location
+from . import test_vertical_lift_tray_type
+from . import test_vertical_lift_shuttle
