+++ stock_vertical_lift/views/stock_location_views.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+  <record id="view_location_form_vertical_lift" model="ir.ui.view">
+    <field name="name">stock.location.form.vertical.lift</field>
+    <field name="model">stock.location</field>
+    <field name="inherit_id" ref="stock.view_location_form"/>
+    <field name="arch" type="xml">
+      <field name="return_location" position="after">
+        <field name="vertical_lift_location"
+               attrs="{'invisible': [('vertical_lift_kind', '!=', False), ('vertical_lift_kind', '!=', 'view')]}"
+               />
+      </field>
+      <group name="localization" position="after">
+        <group string="Vertical Lift" name="vertical_lift"
+               attrs="{'invisible': [('vertical_lift_kind', 'not in', ('shuttle', 'tray', 'cell'))]}">
+          <field name="vertical_lift_kind"/>
+          <field name="vertical_lift_tray_type_id"
+                 attrs="{'invisible': [('vertical_lift_kind', '!=', 'tray')], 'required': [('vertical_lift_kind', '=', 'tray')]}"/>
+          <field name="tray_matrix"
+                 widget="location_tray_matrix"
+                 attrs="{'invisible': [('vertical_lift_kind', 'not in', ('tray', 'cell'))]}"
+                 options="{'click_action': 'action_tray_matrix_click'}"
+                 />
+        </group>
+      </group>
+
+      <field name="posx" position="attributes">
+        <attribute name="attrs">{'readonly': [('vertical_lift_kind', '=', 'cell')]}</attribute>
+      </field>
+      <field name="posy" position="attributes">
+        <attribute name="attrs">{'readonly': [('vertical_lift_kind', '=', 'cell')]}</attribute>
+      </field>
+    </field>
+  </record>
+
+  <record id="view_location_search_vertical_lift" model="ir.ui.view">
+    <field name="name">stock.location.search.vertical.lift</field>
+    <field name="model">stock.location</field>
+    <field name="inherit_id" ref="stock.view_location_search"/>
+    <field name="arch" type="xml">
+      <field name="location_id" position="after">
+        <separator/>
+        <field name="vertical_lift_kind"/>
+        <field name="vertical_lift_tray_type_id"/>
+      </field>
+    </field>
+  </record>
+
+</odoo>
