+++ stock_vertical_lift/views/vertical_lift_tray_type_views.xml
@@ -1,0 +1,85 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+  <record id="view_vertical_lift_tray_type_form" model="ir.ui.view">
+    <field name="name">vertical.lift.tray.type.form</field>
+    <field name="model">vertical.lift.tray.type</field>
+    <field name="arch" type="xml">
+      <form string="Vertical Lift Tray Type">
+        <div class="oe_button_box" name="button_box">
+          <button string="Locations"
+                  class="oe_stat_button"
+                  icon="fa-filter" name="open_locations"
+                  type="object"
+                  />
+          <button name="toggle_active" type="object" class="oe_stat_button" icon="fa-archive">
+            <field name="active" widget="boolean_button" options='{"terminology": "archive"}'/>
+          </button>
+        </div>
+        <label for="name" class="oe_edit_only"/>
+        <h1><field name="name"/></h1>
+        <group>
+          <group string="Tray Configuration" name="settings">
+            <field name="code"/>
+            <field name="rows"/>
+            <field name="cols"/>
+          </group>
+          <group string="Disposition" name="disposition">
+            <field name="tray_matrix"
+                   widget="location_tray_matrix"
+                   nolabel="1"/>
+          </group>
+        </group>
+      </form>
+    </field>
+  </record>
+
+  <record id="view_vertical_lift_tray_type_search" model="ir.ui.view">
+    <field name="name">vertical.lift.tray.type.search</field>
+    <field name="model">vertical.lift.tray.type</field>
+    <field name="arch" type="xml">
+      <search string="Vertical Lift Tray Type">
+        <field name="name"/>
+        <field name="code"/>
+        <separator/>
+        <filter string="Archived" name="inactive" domain="[('active','=',False)]"/>
+      </search>
+    </field>
+  </record>
+
+  <record id="view_vertical_lift_tray_type_tree" model="ir.ui.view">
+    <field name="name">vertical.lift.tray.type</field>
+    <field name="model">vertical.lift.tray.type</field>
+    <field name="arch" type="xml">
+      <tree string="Vertical Lift Tray Type">
+        <field name="name"/>
+        <field name="code"/>
+        <field name="rows"/>
+        <field name="cols"/>
+      </tree>
+    </field>
+  </record>
+
+  <record id="action_vertical_lift_tray_type" model="ir.actions.act_window">
+    <field name="name">Vertical Lift Tray Types</field>
+    <field name="res_model">vertical.lift.tray.type</field>
+    <field name="type">ir.actions.act_window</field>
+    <field name="view_type">form</field>
+    <field name="view_id" ref="view_vertical_lift_tray_type_tree"/>
+    <field name="search_view_id" ref="view_vertical_lift_tray_type_search"/>
+    <field name="context"></field>
+    <field name="help" type="html">
+      <p class="o_view_nocontent_smiling_face">
+        Add a Vertical Lift Tray Type
+        </p><p>
+        Define the number of rows and cols on a tray, depending of the boxes
+size.
+      </p>
+    </field>
+  </record>
+
+  <menuitem action="action_vertical_lift_tray_type"
+            id="menu_vertical_lift_tray_type"
+            parent="stock.menu_warehouse_config" sequence="10"/>
+
+</odoo>
